app/requirements.txt
uvicorn
fastapi
pydantic
httpx

singleton_package
json-advanced
python-dotenv
debugpy
ipython

aiofiles
aiocache

beanie
fastapi-mongo-base
ufaas-fastapi-business

apscheduler
pytz

ufiles
ufaas
usso[fastapi]

ffmpeg